Fix models pkg name
