+++ models/devices.go
@@ -1,4 +1,4 @@
-package model
+package models
 
 type Device struct {
 	Status     string `json:"status"`
